Main.groovy refactor and improvements
DB and Google logging code are separated from Main.groovy to new files
Google spreadsheet URL is printed
Report from CSV comparision is completed
Error report is stored into file in summary results dir
Documentation
